refactor(chat-kotlin): extract MessageControllerTest from ChatController
