+++ challenge-load-test-chat/chat-kotlin/src/main/kotlin/com/example/chatkotlin/handler/ApiExceptionHandler.kt
@@ -1,0 +1,14 @@
+package com.example.chatkotlin.handler
+
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ControllerAdvice
+import org.springframework.web.bind.annotation.ExceptionHandler
+import java.util.NoSuchElementException
+
+@ControllerAdvice
+class ApiExceptionHandler {
+    @ExceptionHandler(NoSuchElementException::class)
+    fun handleNotFound(e: NoSuchElementException): ResponseEntity<Map<String, String>> =
+        ResponseEntity.status(HttpStatus.NOT_FOUND).body(mapOf("error" to (e.message ?: "Not found")))
+}
